requirements.txt
google-adk>=1.5.0
google-genai>=1.0.0
python-dotenv>=1.0.1
chainlit>=2.1.0
pytest>=8.0.0
